+++ prj/vs2022/CvtOnnx.vcxproj.filters
@@ -207,6 +207,9 @@
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertConstantOfShapeNode.cpp">
       <Filter>OnnxRuntime\C</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertPadNode.cpp">
+      <Filter>OnnxRuntime\P</Filter>
+    </ClCompile>
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertPreluNode.cpp">
       <Filter>OnnxRuntime\P</Filter>
     </ClCompile>
