+++ src/Cvt/OnnxRuntime/Convert.h
@@ -135,6 +135,8 @@ namespace Synet
 
     bool ConvertNotNode(const onnx::NodeProto& node, LayerParam& layer);
 
+    bool ConvertPadNode(const onnx::NodeProto& node, const LayerParams& layers, const Bytes& original, LayerParam& layer);
+
     bool ConvertPreluNode(const onnx::NodeProto& node, LayerParams& layers, LayerParam& layer);
 
     bool ConvertQLinearAddNode(const onnx::NodeProto& node, const LayerParams& layers, const Bytes& srcBin, LayerParam& layer);
