+++ src/Cvt/OnnxRuntime/OnnxRuntime.h
@@ -61,50 +61,6 @@ namespace Synet
 
         //-----------------------------------------------------------------------------------------
 
-        bool ConvertPadNode(const onnx::NodeProto& node, const LayerParams& layers, const Bytes& original, LayerParam& layer)
-        {
-            if (!CheckSourceNumber(layer, 1, 3))
-                return false;
-
-            layer.type() = Synet::LayerTypePad;
-            String mode;
-            if (!ConvertAtrributeString(node, "mode", mode, true, "constant"))
-                return false;
-            if (mode == "constant")
-                layer.pad().mode() = PadModeConstant;
-            else if (mode == "reflect")
-                layer.pad().mode() = PadModeReflect;
-            else if (mode == "edge")
-                layer.pad().mode() = PadModeEdge;
-            else if (mode == "wrap")
-                layer.pad().mode() = PadModeWrap;
-            else
-                SYNET_ERROR("Unknown type of pad mode: " << mode << " !");
-            
-            if (layer.src().size() > 1)
-            {
-                const LayerParam* src1 = GetLayer(layers, layer.src()[1]);
-                if (src1 == NULL || src1->type() != LayerTypeMeta)
-                    return false;
-                if (layer.src().size() > 2)
-                {
-                    const LayerParam* src2 = GetLayer(layers, layer.src()[2]);
-                    if (src2 == NULL || src2->type() != LayerTypeConst || src2->weight()[0].type() != TensorType32f)
-                        return false;
-                    if(GetWeight<float>(original, src2->weight()[0])[0] != 0)
-                        SYNET_ERROR("Synet support only pad value == 0!");
-                    layer.src().resize(2);
-                }
-            }  
-            else
-            {
-                if (!ConvertAtrributeInts(node, "pads", layer.pad().pads()))
-                    return false;
-            }
-
-            return true;
-        }
-
         bool ConvertPowNode(const onnx::NodeProto& node, const LayerParams& layers, const Bytes& original, LayerParam& layer)
         {
             if (!CheckSourceNumber(layer, 2))
